Update discogs client folder structure
