+++ src/utils/clients/discogs/index.ts
@@ -1,0 +1,1 @@
+export * from './methods';
