+++ src/utils/clients/discogs/methods/index.ts
@@ -1,0 +1,1 @@
+export * from './album-search';
